test: E2E test case added and Unit test cases added
ALS-1820
