+++ app/src/androidTest/java/com/aws/amazonlocation/ConstantTest.kt
@@ -38,6 +38,7 @@ const val TEST_WORD_KLUANG = "Kluang "
 const val TEST_WORD_ARG = "ARG"
 const val TEST_WORD_RUS = "RUS"
 const val TEST_WORD_ARRIVE = "Arrive"
+const val TEST_WORD_LEAVE = "Leave"
 const val TEST_WORD_LANGUAGE_AR = "العربية"
 const val TEST_WORD_LANGUAGE_BO = "Bosanski"
 const val TEST_FAILED = "Test failed"
